Clabot fix rebase login (#22)
 [FIX] bad handling of authors committers

in some cases (not quite clear) when the committer has no registered github login
we could get a bad handling of the authors and the committers. This PR fixes this problem.

it also refactors a little bit the code to be able to add some unit tests and indeed
adds some tests.